--- helm/helm.go
@@ -1,2 +1,0 @@
-// Package helm contains utility functions for managing and using helm
-package helm
